Move to use Decimals everywhere

+++ care/emr/models/supply_delivery.py
@@ -10,7 +10,9 @@ class SupplyDelivery(EMRBaseModel):
         null=True, blank=True, default=None
     )
     supplied_item_pack_size = models.IntegerField(null=True, blank=True, default=None)
-    supplied_item_quantity = models.FloatField(null=True, blank=True)
+    supplied_item_quantity = models.DecimalField(
+        null=True, blank=True, max_digits=20, decimal_places=6
+    )
     supplied_item = models.ForeignKey(
         "emr.Product", on_delete=models.CASCADE, null=True, blank=True
     )
